MAF/Build_real_world_agents/Product_conceptualizer/requirements.txt
a2a-sdk==0.3.14
google-adk==1.18.0
python-dotenv==1.2.1
uvicorn==0.38.0
fastapi==0.121.2
pydantic==2.12.4
google-generativeai==0.8.5
agent-framework==1.0.0b251114
agent-framework-a2a==1.0.0b251114
agent-framework-azure-ai==1.0.0b251114
agent-framework-core==1.0.0b251114